requirements.txt
Flask>=2.0.1
Flask-CORS>=3.0.10
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
spotipy>=2.19.0
python-dotenv>=0.19.0
